name shared modules according to their benchmark
